requirements.txt
flask==3.0.3
python-dotenv==1.0.1
openai==1.47.0
httpx==0.27.0
